+++ controllers/ajuste.go
@@ -185,3 +185,46 @@ func (c *AjusteController) Put() {
 
 	c.ServeJSON()
 }
+
+// GetOne ...
+// @Title GetOneAuto
+// @Description Retorna la lista de elementos asociados a un acta con su respectiva vida útil y valor residual iniciales
+// @Param	id		path 	string	true		"The key for staticblock"
+// @Success 200 {object} []models.DetalleElemento__
+// @Failure 403 :id is empty
+// @router /automatico/elementos/:id [get]
+func (c *AjusteController) GetOneAuto() {
+
+	defer errorctrl.ErrorControlController(c.Controller, "AjusteController")
+
+	var id int
+	if v, err := c.GetInt(":id"); err != nil || v <= 0 {
+		if err == nil {
+			err = errors.New("Se debe especificar un acta válida")
+		}
+		logs.Error(err)
+		panic(map[string]interface{}{
+			"funcion": `GetOneAuto - c.GetInt(":id")`,
+			"err":     err,
+			"status":  "400",
+		})
+	} else {
+		id = v
+	}
+
+	if respuesta, err := ajustesHelper.GetDetalleElementosActa(id); err == nil || respuesta != nil {
+		c.Data["json"] = respuesta
+	} else {
+		if err != nil {
+			panic(err)
+		}
+
+		panic(map[string]interface{}{
+			"funcion": "GetOneAuto - ajustesHelper.GetDetalleElementosActa(id)",
+			"err":     errors.New("No se obtuvo respuesta al consultar los elementos"),
+			"status":  "404",
+		})
+	}
+
+	c.ServeJSON()
+}
