+++ helpers/ajustesHelper/ajustes.acta.helper.go
@@ -1,0 +1,160 @@
+package ajustesHelper
+
+import (
+	"fmt"
+	"strconv"
+	"time"
+
+	"github.com/astaxie/beego"
+	"github.com/astaxie/beego/logs"
+	"github.com/udistrital/arka_mid/helpers/catalogoElementosHelper"
+	"github.com/udistrital/arka_mid/helpers/utilsHelper"
+	"github.com/udistrital/arka_mid/models"
+	"github.com/udistrital/utils_oas/errorctrl"
+	"github.com/udistrital/utils_oas/formatdata"
+)
+
+// determinarDeltaActa Separa elementos según el ajuste
+func determinarDeltaActa(org *models.Elemento, nvo *models.DetalleElemento_) (msc, vls, sg bool, outputError map[string]interface{}) {
+
+	funcion := "determinarDeltaActa"
+	defer errorctrl.ErrorControlFunction(funcion+" - Unhandled Error!", "500")
+
+	if org.SubgrupoCatalogoId != nvo.SubgrupoCatalogoId {
+
+		urlcrud := "fields=TipoBienId&sortby=Id&order=desc&query=Activo:true,SubgrupoId__Id:" + strconv.Itoa(nvo.SubgrupoCatalogoId)
+		if detalleSubgrupo_, err := catalogoElementosHelper.GetAllDetalleSubgrupo(urlcrud); err != nil {
+			return false, false, false, err
+		} else if len(detalleSubgrupo_) == 0 {
+			err := "len(detalleSubgrupo_) = 0"
+			eval := " - catalogoElementosHelper.GetAllDetalleSubgrupo(urlcrud)"
+			return false, false, false, errorctrl.Error(funcion+eval, err, "500")
+		} else {
+			if detalleSubgrupo_[0].TipoBienId.NecesitaPlaca && nvo.Placa != "" {
+				ctxPlaca, _ := beego.AppConfig.Int("contxtPlaca")
+				if placa_, _, err := utilsHelper.GetConsecutivo("%05.0f", ctxPlaca, "Registro Placa Arka"); err != nil {
+					return false, false, false, err
+				} else {
+					year, month, day := time.Now().Date()
+					nvo.Placa = utilsHelper.FormatConsecutivo(fmt.Sprintf("%04d%02d%02d", year, month, day), placa_, "")
+				}
+			} else if !detalleSubgrupo_[0].TipoBienId.NecesitaPlaca && nvo.Placa != "" {
+				nvo.Placa = ""
+			}
+
+		}
+
+		nvo.Activo = true
+		sg = true
+
+	} else if org.ValorTotal != nvo.ValorTotal {
+		nvo.Activo = true
+		vls = true
+
+	} else if org.Nombre != nvo.Nombre || org.Marca != nvo.Marca ||
+		org.Serie != nvo.Serie || org.UnidadMedida != nvo.UnidadMedida ||
+		org.Cantidad != nvo.Cantidad || org.ValorUnitario != nvo.ValorUnitario ||
+		org.Subtotal != nvo.Subtotal || org.Descuento != nvo.Descuento ||
+		org.PorcentajeIvaId != nvo.PorcentajeIvaId ||
+		org.ValorIva != nvo.ValorIva || org.ValorFinal != nvo.ValorFinal {
+
+		nvo.Activo = true
+		msc = true
+
+	}
+
+	return msc, vls, sg, nil
+
+}
+
+// fillElementos Consulta el detalle de los subgrupos
+func fillElementos(elsOrg []*models.DetalleElemento_) (completos []*models.DetalleElemento__, outputError map[string]interface{}) {
+
+	funcion := "fillElementos"
+	defer errorctrl.ErrorControlFunction(funcion+" - Unhandled Error!", "500")
+
+	var (
+		ids       []int
+		query     string
+		subgrupos map[int]*models.DetalleSubgrupo
+	)
+
+	for _, el := range elsOrg {
+		ids = append(ids, el.SubgrupoCatalogoId)
+	}
+
+	query = "fields=SubgrupoId,TipoBienId,Depreciacion,Amortizacion,ValorResidual,VidaUtil&sortby=Id&order=desc"
+	query += "&query=Activo:true,SubgrupoId__Id__in:" + utilsHelper.ArrayToString(ids, "|")
+	if sg, err := catalogoElementosHelper.GetAllDetalleSubgrupo(query); err != nil {
+		return nil, err
+	} else {
+		subgrupos = make(map[int]*models.DetalleSubgrupo)
+		for _, sg := range sg {
+			if subgrupos[sg.SubgrupoId.Id] == nil {
+				subgrupos[sg.SubgrupoId.Id] = sg
+			}
+		}
+	}
+
+	for _, el := range elsOrg {
+		elC := new(models.DetalleElemento__)
+		elC.Id = 0
+		elC.Nombre = el.Nombre
+		elC.Cantidad = el.Cantidad
+		elC.Marca = el.Marca
+		elC.Serie = el.Serie
+		elC.UnidadMedida = el.UnidadMedida
+		elC.ValorUnitario = el.ValorUnitario
+		elC.Subtotal = el.Subtotal
+		elC.Descuento = el.Descuento
+		elC.ValorTotal = el.ValorTotal
+		elC.PorcentajeIvaId = el.PorcentajeIvaId
+		elC.ValorIva = el.ValorIva
+		elC.ValorFinal = el.ValorFinal
+		elC.SubgrupoCatalogoId = subgrupos[el.SubgrupoCatalogoId]
+		elC.EstadoElementoId = el.EstadoElementoId
+		elC.ActaRecibidoId = el.ActaRecibidoId
+		elC.Placa = el.Placa
+		elC.ValorResidual = el.ValorResidual
+		elC.VidaUtil = el.VidaUtil
+
+		completos = append(completos, elC)
+
+	}
+
+	return completos, nil
+
+}
+
+// fillElemento Agrega la vida útil y valor residual al elemento del acta
+func fillElemento(elActa *models.DetalleElemento, elMov *models.ElementosMovimiento) (completo *models.DetalleElemento__, outputError map[string]interface{}) {
+
+	funcion := "fillElemento"
+	defer errorctrl.ErrorControlFunction(funcion+" - Unhandled Error!", "500")
+
+	if err := formatdata.FillStruct(elActa, &completo); err != nil {
+		logs.Error(err)
+		eval := " - formatdata.FillStruct(elActa, &completo)"
+		return nil, errorctrl.Error(funcion+eval, err, "500")
+	}
+	completo.VidaUtil = elMov.VidaUtil
+	completo.ValorResidual = elMov.ValorResidual
+
+	return completo, nil
+
+}
+
+func generarNuevosActa(nuevos []*models.DetalleElemento_) (actualizados []*models.Elemento, outputError map[string]interface{}) {
+
+	funcion := "generarNuevosActa"
+	defer errorctrl.ErrorControlFunction(funcion+" - Unhandled Error!", "500")
+
+	if err := formatdata.FillStruct(nuevos, &actualizados); err != nil {
+		logs.Error(err)
+		eval := " - formatdata.FillStruct(nuevos, &actualizados)"
+		return nil, errorctrl.Error(funcion+eval, err, "500")
+	}
+
+	return actualizados, nil
+
+}
