+++ swagger/swagger.json
@@ -291,6 +291,38 @@
                 }
             }
         },
+        "/ajustes/automatico/elementos/{id}": {
+            "get": {
+                "tags": [
+                    "ajustes"
+                ],
+                "description": "Retorna la lista de elementos asociados a un acta con su respectiva vida útil y valor residual iniciales",
+                "operationId": "AjusteController.GetOneAuto",
+                "parameters": [
+                    {
+                        "in": "path",
+                        "name": "id",
+                        "description": "The key for staticblock",
+                        "required": true,
+                        "type": "string"
+                    }
+                ],
+                "responses": {
+                    "200": {
+                        "description": "",
+                        "schema": {
+                            "type": "array",
+                            "items": {
+                                "$ref": "#/definitions/models.DetalleElemento__"
+                            }
+                        }
+                    },
+                    "403": {
+                        "description": ":id is empty"
+                    }
+                }
+            }
+        },
         "/ajustes/{id}": {
             "get": {
                 "tags": [
@@ -1760,7 +1792,7 @@
                 "Elementos": {
                     "type": "array",
                     "items": {
-                        "$ref": "#/definitions/models.DetalleElemento_"
+                        "$ref": "#/definitions/models.DetalleElemento__"
                     }
                 },
                 "Movimiento": {
@@ -1844,6 +1876,84 @@
                 },
                 "SubtipoMovimientoId": {
                     "$ref": "#/definitions/models.FormatoTipoMovimiento"
+                }
+            }
+        },
+        "models.DetalleElemento": {
+            "title": "DetalleElemento",
+            "type": "object",
+            "properties": {
+                "ActaRecibidoId": {
+                    "$ref": "#/definitions/models.ActaRecibido"
+                },
+                "Activo": {
+                    "type": "boolean"
+                },
+                "Cantidad": {
+                    "type": "integer",
+                    "format": "int64"
+                },
+                "Descuento": {
+                    "type": "number",
+                    "format": "double"
+                },
+                "EstadoElementoId": {
+                    "$ref": "#/definitions/models.EstadoElemento"
+                },
+                "FechaCreacion": {
+                    "type": "string",
+                    "format": "datetime"
+                },
+                "FechaModificacion": {
+                    "type": "string",
+                    "format": "datetime"
+                },
+                "Id": {
+                    "type": "integer",
+                    "format": "int64"
+                },
+                "Marca": {
+                    "type": "string"
+                },
+                "Nombre": {
+                    "type": "string"
+                },
+                "Placa": {
+                    "type": "string"
+                },
+                "PorcentajeIvaId": {
+                    "type": "integer",
+                    "format": "int64"
+                },
+                "Serie": {
+                    "type": "string"
+                },
+                "SubgrupoCatalogoId": {
+                    "$ref": "#/definitions/models.DetalleSubgrupo"
+                },
+                "Subtotal": {
+                    "type": "number",
+                    "format": "double"
+                },
+                "UnidadMedida": {
+                    "type": "integer",
+                    "format": "int64"
+                },
+                "ValorFinal": {
+                    "type": "number",
+                    "format": "double"
+                },
+                "ValorIva": {
+                    "type": "number",
+                    "format": "double"
+                },
+                "ValorTotal": {
+                    "type": "number",
+                    "format": "double"
+                },
+                "ValorUnitario": {
+                    "type": "number",
+                    "format": "double"
                 }
             }
         },
@@ -1963,6 +2073,92 @@
                 "SubgrupoCatalogoId": {
                     "type": "integer",
                     "format": "int64"
+                },
+                "Subtotal": {
+                    "type": "number",
+                    "format": "double"
+                },
+                "UnidadMedida": {
+                    "type": "integer",
+                    "format": "int64"
+                },
+                "ValorFinal": {
+                    "type": "number",
+                    "format": "double"
+                },
+                "ValorIva": {
+                    "type": "number",
+                    "format": "double"
+                },
+                "ValorResidual": {
+                    "type": "number",
+                    "format": "double"
+                },
+                "ValorTotal": {
+                    "type": "number",
+                    "format": "double"
+                },
+                "ValorUnitario": {
+                    "type": "number",
+                    "format": "double"
+                },
+                "VidaUtil": {
+                    "type": "number",
+                    "format": "double"
+                }
+            }
+        },
+        "models.DetalleElemento__": {
+            "title": "DetalleElemento__",
+            "type": "object",
+            "properties": {
+                "ActaRecibidoId": {
+                    "$ref": "#/definitions/models.ActaRecibido"
+                },
+                "Activo": {
+                    "type": "boolean"
+                },
+                "Cantidad": {
+                    "type": "integer",
+                    "format": "int64"
+                },
+                "Descuento": {
+                    "type": "number",
+                    "format": "double"
+                },
+                "EstadoElementoId": {
+                    "$ref": "#/definitions/models.EstadoElemento"
+                },
+                "FechaCreacion": {
+                    "type": "string",
+                    "format": "datetime"
+                },
+                "FechaModificacion": {
+                    "type": "string",
+                    "format": "datetime"
+                },
+                "Id": {
+                    "type": "integer",
+                    "format": "int64"
+                },
+                "Marca": {
+                    "type": "string"
+                },
+                "Nombre": {
+                    "type": "string"
+                },
+                "Placa": {
+                    "type": "string"
+                },
+                "PorcentajeIvaId": {
+                    "type": "integer",
+                    "format": "int64"
+                },
+                "Serie": {
+                    "type": "string"
+                },
+                "SubgrupoCatalogoId": {
+                    "$ref": "#/definitions/models.DetalleSubgrupo"
                 },
                 "Subtotal": {
                     "type": "number",
