+++ swagger/swagger.yml
@@ -232,6 +232,28 @@ paths:
             $ref: '#/definitions/models.DetalleAjusteAutomatico'
         "403":
           description: body is empty
+  /ajustes/automatico/elementos/{id}:
+    get:
+      tags:
+      - ajustes
+      description: Retorna la lista de elementos asociados a un acta con su respectiva
+        vida útil y valor residual iniciales
+      operationId: AjusteController.GetOneAuto
+      parameters:
+      - in: path
+        name: id
+        description: The key for staticblock
+        required: true
+        type: string
+      responses:
+        "200":
+          description: ""
+          schema:
+            type: array
+            items:
+              $ref: '#/definitions/models.DetalleElemento__'
+        "403":
+          description: :id is empty
   /bajas_elementos/:
     post:
       tags:
@@ -1195,7 +1217,7 @@ definitions:
       Elementos:
         type: array
         items:
-          $ref: '#/definitions/models.DetalleElemento_'
+          $ref: '#/definitions/models.DetalleElemento__'
       Movimiento:
         $ref: '#/definitions/models.Movimiento'
       TrContable:
@@ -1252,6 +1274,62 @@ definitions:
         format: int64
       SubtipoMovimientoId:
         $ref: '#/definitions/models.FormatoTipoMovimiento'
+  models.DetalleElemento:
+    title: DetalleElemento
+    type: object
+    properties:
+      ActaRecibidoId:
+        $ref: '#/definitions/models.ActaRecibido'
+      Activo:
+        type: boolean
+      Cantidad:
+        type: integer
+        format: int64
+      Descuento:
+        type: number
+        format: double
+      EstadoElementoId:
+        $ref: '#/definitions/models.EstadoElemento'
+      FechaCreacion:
+        type: string
+        format: datetime
+      FechaModificacion:
+        type: string
+        format: datetime
+      Id:
+        type: integer
+        format: int64
+      Marca:
+        type: string
+      Nombre:
+        type: string
+      Placa:
+        type: string
+      PorcentajeIvaId:
+        type: integer
+        format: int64
+      Serie:
+        type: string
+      SubgrupoCatalogoId:
+        $ref: '#/definitions/models.DetalleSubgrupo'
+      Subtotal:
+        type: number
+        format: double
+      UnidadMedida:
+        type: integer
+        format: int64
+      ValorFinal:
+        type: number
+        format: double
+      ValorIva:
+        type: number
+        format: double
+      ValorTotal:
+        type: number
+        format: double
+      ValorUnitario:
+        type: number
+        format: double
   models.DetalleElemento_:
     title: DetalleElemento_
     type: object
@@ -1291,6 +1369,68 @@ definitions:
       SubgrupoCatalogoId:
         type: integer
         format: int64
+      Subtotal:
+        type: number
+        format: double
+      UnidadMedida:
+        type: integer
+        format: int64
+      ValorFinal:
+        type: number
+        format: double
+      ValorIva:
+        type: number
+        format: double
+      ValorResidual:
+        type: number
+        format: double
+      ValorTotal:
+        type: number
+        format: double
+      ValorUnitario:
+        type: number
+        format: double
+      VidaUtil:
+        type: number
+        format: double
+  models.DetalleElemento__:
+    title: DetalleElemento__
+    type: object
+    properties:
+      ActaRecibidoId:
+        $ref: '#/definitions/models.ActaRecibido'
+      Activo:
+        type: boolean
+      Cantidad:
+        type: integer
+        format: int64
+      Descuento:
+        type: number
+        format: double
+      EstadoElementoId:
+        $ref: '#/definitions/models.EstadoElemento'
+      FechaCreacion:
+        type: string
+        format: datetime
+      FechaModificacion:
+        type: string
+        format: datetime
+      Id:
+        type: integer
+        format: int64
+      Marca:
+        type: string
+      Nombre:
+        type: string
+      Placa:
+        type: string
+      PorcentajeIvaId:
+        type: integer
+        format: int64
+      Serie:
+        type: string
+      SubgrupoCatalogoId:
+        $ref: '#/definitions/models.DetalleSubgrupo'
       Subtotal:
         type: number
         format: double
